+++ upload/admin/controller/catalog/manufacturer.php
@@ -131,7 +131,7 @@ class Manufacturer extends Controller {
 		$url = $filter->getQueryString(true, true);
 
 		// Total Manufacturers
-		$manufacturer_total = $this->model_catalog_manufacturer->getTotalManufacturers();
+		$manufacturer_total = $this->model_catalog_manufacturer->getTotalManufacturers($filter_data);
 
 		// Pagination
 		$data['pagination'] = $this->load->controller('common/pagination', [
@@ -381,35 +381,12 @@ class Manufacturer extends Controller {
 	}
 
 	public function autocomplete(): void {
-		$json = [];
+		$this->load->model('catalog/manufacturer');
 
-		// Manufacturers
-		if (isset($this->request->get['filter_name'])) {
-			$this->load->model('catalog/manufacturer');
-
-			$filter_data = [
-				'filter_name' => $this->request->get['filter_name'],
-				'start'       => 0,
-				'limit'       => $this->config->get('config_autocomplete_limit')
-			];
-
-			$results = $this->model_catalog_manufacturer->getManufacturers($filter_data);
-
-			foreach ($results as $result) {
-				$json[] = [
-					'manufacturer_id' => $result['manufacturer_id'],
-					'name'            => strip_tags(html_entity_decode($result['name'], ENT_QUOTES, 'UTF-8'))
-				];
-			}
-		}
-
-		$sort_order = [];
-
-		foreach ($json as $key => $value) {
-			$sort_order[$key] = $value['name'];
-		}
-
-		array_multisort($sort_order, SORT_ASC, $json);
+		$json = $this->model_catalog_manufacturer->getManufacturers([
+			'filter_name' => $this->request->get['autocomplete_name'] ?? '',
+			'limit'       => $this->config->get('config_autocomplete_limit')
+		]);
 
 		$this->response->addHeader('Content-Type: application/json');
 		$this->response->setOutput(json_encode($json));