+++ upload/admin/controller/user/user.php
@@ -517,21 +517,21 @@ class User extends Controller {
 	public function autocomplete(): void {
 		$json = [];
 
-		if (isset($this->request->get['filter_username']) || isset($this->request->get['filter_name']) || isset($this->request->get['filter_email'])) {
-			if (isset($this->request->get['filter_username'])) {
-				$filter_username = $this->request->get['filter_username'];
+		if (isset($this->request->get['autocomplete_username']) || isset($this->request->get['autocomplete_name']) || isset($this->request->get['autocomplete_email'])) {
+			if (isset($this->request->get['autocomplete_username'])) {
+				$filter_username = $this->request->get['autocomplete_username'];
 			} else {
 				$filter_username = '';
 			}
 
-			if (isset($this->request->get['filter_name'])) {
-				$filter_name = $this->request->get['filter_name'];
+			if (isset($this->request->get['autocomplete_name'])) {
+				$filter_name = $this->request->get['autocomplete_name'];
 			} else {
 				$filter_name = '';
 			}
 
-			if (isset($this->request->get['filter_email'])) {
-				$filter_email = $this->request->get['filter_email'];
+			if (isset($this->request->get['autocomplete_email'])) {
+				$filter_email = $this->request->get['autocomplete_email'];
 			} else {
 				$filter_email = '';
 			}