+++ upload/admin/model/design/seo_url.php
@@ -433,7 +433,8 @@ class SeoUrl extends \Opencart\System\Engine\Model {
 			$sql .= " GROUP BY `value` ORDER BY `value`";
 		}
 
-		$sql .= " LIMIT " . (int)$data['limit'];
+		$sql .= " LIMIT " . (int)$this->config->get('config_autocomplete_limit');
+
 		$query = $this->db->query($sql);
 
 		return $query->rows;